don't need modules if we're using base class extensions
